+++ buildSrc/src/main/java/Dependencies.kt
@@ -1,22 +1,20 @@
 object Dependencies {
 
     // Gradle
-    const val gradle = "com.android.tools.build:gradle:${Versions.gradle}"
+    const val gradleAndroid = "com.android.tools.build:gradle:${Versions.gradleAndroid}"
     const val gradleKotlin = "org.jetbrains.kotlin:kotlin-gradle-plugin:${Versions.gradleKotlin}"
 
     // Deploy
-    const val bintray = "com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.4"
-    const val maven = "com.github.dcendents:android-maven-gradle-plugin:2.1"
+    const val maven = "com.github.dcendents:android-maven-gradle-plugin:${Versions.mavenPublish}"
 
     // Kotlin
     const val kotlinStdlib = "org.jetbrains.kotlin:kotlin-stdlib-jdk7:${Versions.gradleKotlin}"
 
     // AppCompat
-    const val appCompat = "com.android.support:appcompat-v7:${Versions.appcompat}"
+    const val appCompat = "androidx.appcompat:appcompat:${Versions.appcompat}"
 
     // ConstraintLayout
-    const val constraintLayout =
-        "com.android.support.constraint:constraint-layout:${Versions.constraintLayout}"
+    const val constraintLayout = "androidx.constraintlayout:constraintlayout:${Versions.constraintLayout}"
 
     // Picasso
     const val picasso = "com.squareup.picasso:picasso:${Versions.picasso}"