+++ buildSrc/src/main/java/Versions.kt
@@ -1,16 +1,17 @@
 object Versions {
 
-    const val versionToDeploy = "1.1.1"
+    const val libraryVersion = "1.2.0"
 
-    const val compileSdkVersion = 28
-    const val minSdkVersion = 19
-    const val targetSdkVersion = 28
+    const val minSdkVersion = 21
+    const val sdkVersion = 30
+    const val buildToolsVersion = "30.0.2"
 
-    const val constraintLayout = "1.1.3"
-    const val appcompat = "28.0.0"
-    const val gradle = "3.5.1"
-    const val gradleKotlin = "1.3.50"
-    const val okHttp = "3.11.0"
+    const val mavenPublish = "2.1"
+    const val constraintLayout = "2.0.4"
+    const val appcompat = "1.2.0"
+    const val gradleAndroid = "4.2.1"
+    const val gradleKotlin = "1.3.71"
+    const val okHttp = "3.14.9"
     const val picasso = "2.5.2"
     const val picassoTestApp = "2.71828"
 }